backend/content_service/requirements.txt
fastapi>=0.99.0
uvicorn>=0.22.0
pydantic>=2.5.1
pydantic-settings>=2.3.0
python-multipart>=0.0.6
python-jose>=3.4.0
motor==3.7.1
pytest>=7.4.0
httpx
azure-storage-blob>=12.14.1
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.0
asyncio>=3.4.3
fastapi-mail>=1.4.1
jinja2>=3.1.2
